+++ RELEASENOTES.md
@@ -11,6 +11,7 @@ Just run `grasp_install` to install the missing 3rd party dependencies (see belo
 ### New Third Party Toolboxes
 
  * `bgirault_molene_dataset`: Molène dataset (weather stations in Britanny, France)
+ * `chebfun`: [Chebyshev polynomials toolbox for matlab](https://www.chebfun.org/)
 
 ### New Functions
 
@@ -19,6 +20,7 @@ Just run `grasp_install` to install the missing 3rd party dependencies (see belo
  * `grasp_random_bipartite`: Random bipartite graph
  * `grasp_layout_spectral`: embedding from the random walk Laplacian eigenvectors
  * `grasp_maximally_connected_subgraphs`: computes the maximally connected subgraphs
+ * `grasp_filter_cheb`: creates a Chebyshev polynomial graph filter
 
 ### Minor Features